requirements.txt
Flask
Pillow
cryptography
numpy
opencv-python
